+++ examples/create_build_run_actor/main.go
@@ -18,7 +18,7 @@ import (
 )
 
 func main() {
-	client := apify.NewClient(os.Getenv("APIFY_TOKEN"))
+	client := apify.NewClient(apify.WithToken(os.Getenv("APIFY_TOKEN")))
 	ctx := context.Background()
 
 	name := fmt.Sprintf("go-example-%d", time.Now().Unix())